+++ Assets/01_Scripts/Managers/GameActive.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b434078e83afb2e408e91855d8d897d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
