+++ Assets/01_Scripts/Managers/SaveLoad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87498be3688d75c44a63ed8957eb9106
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
